Fix remaining Float/CGFloat: normalize closedness to CGFloat at the source

+++ modules/camera-engine/ios/CameraEngineModule.swift
@@ -1920,9 +1920,9 @@ enum ApertureSimulationProcessor {
     //    closedness = clamp((f - min) / (max - min), 0, 1); blur = 1 - closedness.
     //    (Wide open -> strongest blur; smallest stop -> weakest. Max 28px.)
     let spec = Camera18ApertureSpec.load()
-    let closedness = min(1.0, max(0.0, (fNumber - spec.min) / (spec.max - spec.min)))
+    let closedness = CGFloat(min(1.0, max(0.0, (fNumber - spec.min) / (spec.max - spec.min))))
     let blurStrength = 1.0 - closedness
-    let radius = blurStrength * CGFloat(28.0)
+    let radius = blurStrength * 28.0
     if radius >= 1.5, let mask = personMask(for: input) {
       // Blur a QUARTER-SCALE copy and upscale: the background is defocused anyway, so
       // the detail loss is invisible, and a 24MP r=24 gaussian stays cheap.
@@ -1988,9 +1988,9 @@ enum ApertureSimulationProcessor {
 enum StarburstProcessor {
   static func apply(_ input: CIImage, fNumber: Float) -> CIImage {
     let spec = Camera18ApertureSpec.load()
-    let closedness = min(1.0, max(0.0, (fNumber - spec.min) / (spec.max - spec.min)))
+    let closedness = CGFloat(min(1.0, max(0.0, (fNumber - spec.min) / (spec.max - spec.min))))
     let strength = closedness
-    let rayRadius = 10.0 + closedness * CGFloat(30.0)
+    let rayRadius = 10.0 + closedness * 30.0
     guard strength > 0.01 else { return input }
     return input.appStarburst(strength: strength, rayRadius: rayRadius)
   }
